+++ pkg/scanner/dom/patterns.go
@@ -62,4 +62,7 @@ var sinkPatterns = []string{
 	`\$sce\.trustAsHtml\(`,
 	// React Sinks
 	`dangerouslySetInnerHTML`,
+	// Sanitization / mXSS
+	`DOMPurify\.sanitize\(`,
+	`sanitizeHtml\(`,
 }